+++ app/src/main/java/org/stepic/droid/ui/fragments/SectionsFragment.java
@@ -52,14 +52,14 @@ import org.stepic.droid.core.presenters.CourseFinderPresenter;
 import org.stepic.droid.core.presenters.CourseJoinerPresenter;
 import org.stepic.droid.core.presenters.DownloadingInteractionPresenter;
 import org.stepic.droid.core.presenters.InvitationPresenter;
-import org.stepic.droid.core.presenters.PersonalDeadlinesPresenter;
+import org.stepic.droid.features.deadlines.presenters.PersonalDeadlinesPresenter;
 import org.stepic.droid.core.presenters.SectionsPresenter;
 import org.stepic.droid.core.presenters.contracts.CalendarExportableView;
 import org.stepic.droid.core.presenters.contracts.CourseJoinView;
 import org.stepic.droid.core.presenters.contracts.DownloadingInteractionView;
 import org.stepic.droid.core.presenters.contracts.InvitationView;
 import org.stepic.droid.core.presenters.contracts.LoadCourseView;
-import org.stepic.droid.core.presenters.contracts.PersonalDeadlinesView;
+import org.stepic.droid.features.deadlines.presenters.contracts.PersonalDeadlinesView;
 import org.stepic.droid.core.presenters.contracts.SectionsView;
 import org.stepic.droid.model.CalendarItem;
 import org.stepic.droid.model.Course;
